requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
distro==1.9.0
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
packaging==24.1
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pytesseract==0.3.13
sniffio==1.3.1
typing_extensions==4.12.2
